==> scripts/missing/real/8_final_plots.Rmd <==
---
title: "Plots for Missing Paper"
author: "Ben Cooper"
date: "21/05/2021"
output: html_document
---

```{r setup and packages, include=FALSE}
require(dplyr)
require(magrittr)
require(readr)
require(tibble)
require(foreach)
require(stringr)
require(doParallel)
require(tools)
require(tidyr)
require(ROCR)
require(ResourceSelection)
require(ggplot2)
require(scales)
require(car)
require(RColorBrewer)
pal <- c(brewer.pal(8, "Set1")[-6], "black")
require(bayestestR)
require(cowplot)
require(xtable)
require(forcats)
require(patchwork)
require(janitor)

source("functions/inverse_logit.R")

calibration <- function(.df)
{
  .df %>% 
    group_by(decile) %>%
    summarise(N = length(mortality),
              observed = (sum(mortality) / length(mortality)) * 100,
              predicted = mean(probs * 100),
              error = sd(probs * 100))
}

```

```{r Load full data}
apache_df <- read_csv("data/apache_real_missing.csv")
apache_scores_inunit <- apache_df %>% 
  select(-row_id, -adm_id,
         -mort_inunit, -age,
         -chronic, -fractioninspiredoxygen,
         -apache_II, -missing_abg, -acute_renal_failure)
apache_additional_inunit <- apache_df %>% 
  select(row_id, adm_id,
         mort_inunit, age,
         chronic, fractioninspiredoxygen,
         apache_II, missing_abg, acute_renal_failure)
rm(apache_df)

apache_df <- read_csv("data/apache_discharge_missing.csv")
apache_scores_30d <- apache_df %>% 
  select(-row_id, -adm_id,
         -mort_30, -age,
         -chronic, -fractioninspiredoxygen,
         -apache_II, -missing_abg, -acute_renal_failure)
apache_additional_30d <- apache_df %>% 
  select(row_id, adm_id,
         mort_30, age,
         chronic, fractioninspiredoxygen,
         apache_II, missing_abg, acute_renal_failure)
rm(apache_df)
```

```{r Descriptive results}
# N in each dataset
length(unique(apache_additional_inunit$adm_id))
length(unique(apache_additional_30d$adm_id))

# Deaths in each dataset
tabyl(apache_additional_inunit$mort_inunit)
tabyl(apache_additional_30d$mort_30)

# Missingness in each dataset
tabyl(is.na(apache_additional_inunit$apache_II))
tabyl(is.na(apache_additional_30d$apache_II))

# Median APACHE-II
apache_additional_inunit %>% 
  filter(!is.na(apache_II)) %>% 
  summarise(median = median(apache_II),
            LQ = quantile(apache_II, 0.25),
            UQ = quantile(apache_II, 0.75))

apache_additional_30d %>% 
  filter(!is.na(apache_II)) %>% 
  summarise(median = median(apache_II),
            LQ = quantile(apache_II, 0.25),
            UQ = quantile(apache_II, 0.75))

```

```{r Discrimination and calibration of complete cases}
# Generate complete cases models
full_model_inunit <- glm(mort_inunit ~ apache_II,
                         data = apache_additional_inunit,
                         family = "binomial")

full_model_30d <- glm(mort_30 ~ apache_II,
                         data = apache_additional_30d,
                         family = "binomial")

# Discrimination----

# Generate probabilities
data.frame(probs = predict(full_model_inunit, 
                           newdata = full_model_inunit$data),
           outcome = full_model_inunit$data$mort_inunit) %>% 
  mutate(probs = inverse_logit(probs)) %>% 
  na.omit() -> probs_mort_inunit

data.frame(probs = predict(full_model_30d, 
                           newdata = full_model_30d$data),
           outcome = full_model_30d$data$mort_30) %>% 
  mutate(probs = inverse_logit(probs)) %>% 
  na.omit() -> probs_mort_30

# Build prediction objects
prediction_mort_inunit <- prediction(probs_mort_inunit$probs,
                                               probs_mort_inunit$outcome)

prediction_mort_30 <- prediction(probs_mort_30$probs,
                                               probs_mort_30$outcome)

# Generate AUC
performance(prediction_mort_inunit, measure = "auc")@y.values[[1]]
performance(prediction_mort_30, measure = "auc")@y.values[[1]]

# Generate performance objects
performance_mort_inunit <- performance(prediction_mort_inunit, "tpr", "fpr")
performance_mort_30 <- performance(prediction_mort_30, "tpr", "fpr")

# Plot
data.frame(x = performance_mort_inunit@x.values[[1]],
           y = performance_mort_inunit@y.values[[1]],
           model = "In-unit mortality") %>% 
  rbind(
    data.frame(x = performance_mort_30@x.values[[1]],
               y = performance_mort_30@y.values[[1]],
               model = "30-day mortality")
  ) %>% 
  ggplot(aes(x, y, colour = model))+
  geom_abline(slope = 1, intercept = 0,
              linetype = "dotted",
              size = 1)+
  geom_path(size = 1, alpha = 0.9)+
  labs(x = "1 - Specificity",
       y = "Sensitivity")+
  theme_classic(20)+
  scale_color_brewer(palette = "Set1",
                     name = "")+
  theme(legend.position = "top") -> A

# Calibration----

# Split data into deciles
deciles_mort_inunit <- tibble(
  mortality = probs_mort_inunit$outcome,
  probs = probs_mort_inunit$probs,
  decile = ntile(probs_mort_inunit$probs, 10)
)

deciles_mort_30 <- tibble(
  mortality = probs_mort_30$outcome,
  probs = probs_mort_30$probs,
  decile = ntile(probs_mort_30$probs, 10)
)

# Calculate calibration
calib_mort_inunit <- calibration(deciles_mort_inunit)
calib_mort_30 <- calibration(deciles_mort_30)

# Quantify calibration
hoslem.test(probs_mort_inunit$outcome,
            probs_mort_inunit$probs, g = 10)
hoslem.test(probs_mort_30$outcome,
            probs_mort_30$probs, g = 10)

# Plot calibration
rbind(calib_mort_inunit %>% mutate(model = "In-unit mortality"), 
      calib_mort_30 %>% mutate(model = "30-day mortality")) %>% 
  ggplot(aes(predicted, observed ,
             colour = model))+
    geom_path(size = 1, alpha = 0.9)+
  geom_pointrange(aes(ymin = observed - error ,
                      ymax = observed + error,
                      y = observed,
                      x = predicted,
                      colour = model),
                  alpha = 0.9) +
    geom_abline(slope = 1, intercept = 0,
              size = 1, linetype = "dotted")+
  theme_classic(20)+
  theme(legend.position = "top")+
  labs(x = "Predicted mortality",
       y = "Observed mortality")+
  scale_colour_brewer(palette = "Set1",
                      name = "") -> B

# Combine & write----
plot_grid(A,B, labels = "AUTO",
          label_size = 32)
ggsave("drafts/missing/figures/calib_discrim.png",
       height = 6, width = 12)
```

```{r Bars of N missing & each variable}

# Correct fio2
apache_additional_inunit$fractioninspiredoxygen[is.na(apache_additional_inunit$fractioninspiredoxygen)] <- 21
apache_additional_30d$fractioninspiredoxygen[is.na(apache_additional_30d$fractioninspiredoxygen)] <- 21

# Create missingness data frames
missing_inunit <-
  data.frame(
    temperature = is.na(apache_scores_inunit$temperature),
    map = is.na(apache_scores_inunit$systolicbp) |
      is.na(apache_scores_inunit$diastolicbp),
    pulse = is.na(apache_scores_inunit$pulse),
    respiratory = is.na(apache_scores_inunit$respiratory),
    sodium = is.na(apache_scores_inunit$sodium),
    potassium = is.na(apache_scores_inunit$potassium),
    creatinine = is.na(apache_scores_inunit$creatinine),
    haematocrit = is.na(apache_scores_inunit$haematocrit),
    whitebloodcount = is.na(apache_scores_inunit$whitebloodcount),
    gcs = is.na(apache_scores_inunit$glasgowcomaeye) |
      is.na(apache_scores_inunit$glasgowcomamotor) |
      is.na(apache_scores_inunit$glasgowcomaverbal),
    arterialpH = is.na(apache_scores_inunit$arterialpH) &
      is.na(apache_scores_inunit$bicarbonate),
    oxygenation = ifelse(
      apache_additional_inunit$fractioninspiredoxygen > 50,
      is.na(apache_scores_inunit$arterialoxygen) |
        is.na(apache_scores_inunit$arterialcarbon),
      is.na(apache_scores_inunit$arterialoxygen)
    ) &
      is.na(apache_scores_inunit$bicarbonate)
  )

missing_30d <-
  data.frame(
    temperature = is.na(apache_scores_30d$temperature),
    map = is.na(apache_scores_30d$systolicbp) |
      is.na(apache_scores_30d$diastolicbp),
    pulse = is.na(apache_scores_30d$pulse),
    respiratory = is.na(apache_scores_30d$respiratory),
    sodium = is.na(apache_scores_30d$sodium),
    potassium = is.na(apache_scores_30d$potassium),
    creatinine = is.na(apache_scores_30d$creatinine),
    haematocrit = is.na(apache_scores_30d$haematocrit),
    whitebloodcount = is.na(apache_scores_30d$whitebloodcount),
    gcs = is.na(apache_scores_30d$glasgowcomaeye) |
      is.na(apache_scores_30d$glasgowcomamotor) |
      is.na(apache_scores_30d$glasgowcomaverbal),
    arterialpH = is.na(apache_scores_30d$arterialpH) &
      is.na(apache_scores_30d$bicarbonate),
    oxygenation = ifelse(
      apache_additional_30d$fractioninspiredoxygen > 50,
      is.na(apache_scores_30d$arterialoxygen) |
        is.na(apache_scores_30d$arterialcarbon),
      is.na(apache_scores_30d$arterialoxygen)
    ) &
      is.na(apache_scores_30d$bicarbonate)
  )

# Frequency of missingess by variable
miss_table <- colMeans(missing_inunit) %>%
  as.data.frame() %>% 
  mutate(model = "In-unit mortality") %>% 
  rbind(
    colMeans(missing_30d) %>% 
      as.data.frame() %>% 
      mutate(model = "30-day mortality")) 
names(miss_table)[1] <- "Freq"
miss_table$varname <- rep(names(missing_30d), 2)

miss_table %>% 
  mutate(Freq = signif(Freq * 100, 3)) %>% 
  pivot_wider(names_from = "model",
              values_from = "Freq") %>% 
  relocate(varname, `30-day mortality`) %>% 
  xtable() %>% 
  print(include.rownames=FALSE)

  
# Plot
miss_table %>% 
ggplot(aes(x = varname, y = Freq * 100))+
  geom_bar(aes(fill = model),
           stat = "identity",
           position = "dodge",
           colour = "black",
           alpha = 0.8) +
  theme_classic(20)+
  theme(legend.position = "none")+
  labs(x = "",
       y = "Missing rate (%)")+
  scale_fill_brewer(palette = "Set1",
                    name = "")+
  coord_flip()+
  facet_wrap(~model, scales = "free")

# Number of variables missing
missN <- table(rowSums(missing_inunit)) %>% 
  as.data.frame() %>% 
  mutate(model = "In-unit mortality") %>% 
  rbind(table(rowSums(missing_30d)) %>% 
          as.data.frame() %>% 
          mutate(model = "30-day mortality")) %>%
  filter(Var1 != 0)

nvar_table <- missN %>% 
  pivot_wider(names_from = "model",
              values_from = "Freq") %>% 
  mutate(`In-unit mortality` = signif((`In-unit mortality` / nrow(apache_additional_inunit)) * 100, 3),
         `30-day mortality` = signif((`30-day mortality` / nrow(apache_additional_30d)) * 100, 3)) %>% 
  relocate(Var1, `30-day mortality`) %>% 
  xtable() 

digits(nvar_table) <- c(0,0,3,3)
nvar_table %>% 
  print(include.rownames=FALSE)

# Plot
missN %>% 
  ggplot(aes(x = Var1, y = Freq))+
  geom_bar(aes(fill = model),
           stat = "identity",
           position = "dodge",
           colour = "black",
           alpha = 0.8) +
  theme_classic(20)+
  theme(legend.position = "top")+
  labs(x = "Number of variables missing",
       y = "Frequency")+
  scale_fill_brewer(palette = "Set1",
                    name = "")+
  coord_flip()+
  facet_wrap(~model, scales = "free")
```

```{r Bootstrap results - in-unit}
# Load data
boot_samples_inunit <- 
  read_rds("models/boot_samples.RDS") %>% 
  filter(mortality == "inunit")

# Calculate means
means_df <- boot_samples_inunit %>% 
  group_by(method) %>% 
  summarise(discrimination = median(discrim),
            calibration = median(calib))

boot_samples_inunit %>% 
  ggplot(aes(x = discrim, y = calib,
                   colour = method))+
  theme_classic(20)+
  labs(y = "Calibration", x = "Discrimination")+
  scale_colour_manual(values = pal[-6],
                      name = "")+
  scale_fill_manual(values = pal[-6],
                    name = "")+
  scale_y_reverse()+
  theme(legend.position = "top")+
  stat_ellipse(size = 2, type = "norm",
               level = 0.682,
               alpha = 0.8)+
  geom_point(data = means_df,
             aes(x = discrimination,
                 y = calibration,
                 fill = method),
             size = 4, shape = 21)+
  facet_wrap(~method, nrow = 2)

ggsave("drafts/missing/figures/boot_inunit.png",
       height = 7, width = 12, scale = 1.25)
```

```{r Bootstrap results - 30-day}
# Load data
boot_samples_30d <- 
  read_rds("models/boot_samples.RDS") %>% 
  filter(mortality == "30-day")

# Calculate means
means_df <- boot_samples_30d %>% 
  group_by(method) %>% 
  summarise(discrimination = median(discrim),
            calibration = median(calib))

boot_samples_30d %>% 
  ggplot(aes(x = discrim, y = calib,
                   colour = method))+
  theme_classic(20)+
  labs(y = "Calibration", x = "Discrimination")+
  scale_colour_manual(values = pal,
                      name = "")+
  scale_fill_manual(values = pal,
                    name = "")+
  scale_y_reverse()+
  theme(legend.position = "top")+
  stat_ellipse(size = 2, type = "norm",
               level = 0.682,
               alpha = 0.8)+
  geom_point(data = means_df,
             aes(x = discrimination,
                 y = calibration,
                 fill = method),
             size = 4, shape = 21)+
  facet_wrap(~method, nrow = 2)

ggsave("drafts/missing/figures/boot_30d.png",
       height = 7, width = 12, scale = 1.25)

```

```{r Table of HDIs}
# In-unit
HDI_inunit <- boot_samples_inunit %>% 
  mutate(discrim = round(discrim, 3),
         calib = round(calib, 2)) %>% 
  pivot_longer(2:3) %>% 
  group_by(method, name) %>%
  summarise(median = median(value),
    HDI = hdi(value, 0.95)) %>% 
  ungroup() 

# 30-day
HDI_30d <- boot_samples_30d %>% 
    mutate(discrim = round(discrim, 3),
         calib = round(calib, 2)) %>% 
  pivot_longer(2:3) %>% 
  group_by(method, name) %>% 
  summarise(median = median(value),
    HDI = hdi(value, 0.95)) %>% 
  ungroup() 

# Combine
tibble(HDI_inunit[,1:3], CI_low = HDI_inunit$HDI$CI_low,
       CI_high = HDI_inunit$HDI$CI_high, mortality = "In-unit") %>% 
  rbind(data.frame(HDI_30d[,1:3], CI_low = HDI_30d$HDI$CI_low,
       CI_high = HDI_30d$HDI$CI_high, mortality = "30-day")) %>% 
  pivot_wider(names_from = "name",
              values_from = c("median", "CI_low",
                              "CI_high")) %>% 
  mutate(calibration = paste(median_calib, " [", CI_low_calib, 
                             ", ", CI_high_calib, "]", sep = ""),
         discrimination = paste(median_discrim, " [", CI_low_discrim, 
                             ", ", CI_high_discrim, "]", sep = "")) %>% 
  select(-median_calib, -CI_low_calib, -CI_high_calib,
         -median_discrim, -CI_low_discrim, -CI_high_discrim) %>% 
  arrange(mortality, method) %>% 
  relocate(mortality) %>% 
  xtable() %>% 
  print(include.rownames=FALSE)

```

```{r Subsets and hybrids}
# Load data
results_subset <- read_rds("models/boot_results_subset.RDS") 

# Adjust factor levels
results_subset %<>% 
  pivot_longer(2:7) %>% 
  mutate(measure = str_split(name, "_", 2, TRUE)[,1],
         type = str_split(name, "_", 2, TRUE)[,2]) %>% 
  select(-name) %>% 
  pivot_wider(names_from = measure,
              values_from = value) %>% 
  mutate(type = fct_recode(type, Amelia = "amelia",
                           MICE = "mice",
                           Longitudinal = "long"))

# Extract means
means_df_subset <- results_subset %>% 
  group_by(type) %>% 
  summarise(discrimination = mean(discrim),
            calibration = mean(calib)) %>% 
  ungroup()


# Plot
A <- results_subset %>% 
  # filter(mortality == "30-day mortality",
  #        method == "Assume zero") %>% 
  ggplot()+
  theme_classic(20)+
  labs(y = "Calibration", x = "Discrimination")+
  scale_colour_manual(values = c("#e41a1c", "#a65628", 
                                 "#984ea3"), name = "")+
  scale_fill_manual(values = c("#e41a1c", "#a65628", 
                               "#984ea3"), name = "")+
  scale_y_reverse()+
  theme(legend.position = "top")+
  stat_ellipse(aes(x = discrim, y = calib,
                   colour = type),
               size = 2, type = "norm",
               level = 0.682,
               alpha = 0.8)+
  geom_point(data = means_df_subset ,
             aes(x = discrimination,
                 y = calibration,
                 fill = type),
             size = 4, shape = 21)+
  coord_cartesian(xlim = c(0.73, 0.9),
                  ylim = c(70, 5))
  

# Load data
results_hybrid <- read_rds("models/boot_results_hybrid.RDS") %>% 
  as_tibble()

# Adjust factor levels
results_hybrid %<>% 
  pivot_longer(2:5) %>% 
  mutate(measure = str_split(name, "_", 2, TRUE)[,1],
         type = str_split(name, "_", 2, TRUE)[,2]) %>% 
  select(-name) %>% 
  pivot_wider(names_from = measure,
              values_from = value) %>% 
  mutate(type = fct_recode(type, Amelia = "amelia",
                           MICE = "mice"))

# Extract means
means_df_hybrid <- results_hybrid %>% 
  group_by(type) %>% 
  summarise(discrimination = mean(discrim),
            calibration = mean(calib)) %>% 
  ungroup()


# Plot
B <- results_hybrid %>% 
  # filter(mortality == "30-day mortality",
  #        method == "Assume zero") %>% 
  ggplot()+
  theme_classic(20)+
  labs(y = "Calibration", x = "Discrimination")+
  scale_colour_manual(values = c("#e41a1c",
                                 "#984ea3"), name = "")+
  scale_fill_manual(values = c("#e41a1c", 
                               "#984ea3"), name = "")+
  scale_y_reverse()+
  theme(legend.position = "top")+
  stat_ellipse(aes(x = discrim, y = calib,
                   colour = type),
               size = 2, type = "norm",
               level = 0.682,
               alpha = 0.8)+
  geom_point(data = means_df_hybrid ,
             aes(x = discrimination,
                 y = calibration,
                 fill = type),
             size = 4, shape = 21)+
  coord_cartesian(xlim = c(0.73, 0.9),
                  ylim = c(70, 5))
# Combine
A + B
ggsave("drafts/missing/figures/subset_hybrid.png",
       height = 5, width = 10, scale = 1.25)

# Grayscale
boot_samples_30d %>% 
  ggplot(aes(x = discrim, y = calib,
                   colour = method))+
  theme_classic(20)+
  labs(y = "Calibration", x = "Discrimination")+
  scale_y_reverse()+
  theme(legend.position = "top")+
  stat_ellipse(size = 2, type = "norm",
               level = 0.682,
               alpha = 0.8,)+
  geom_point(data = means_df,
             aes(x = discrimination,
                 y = calibration,
                 fill = method),
             size = 4, shape = 21)+
  coord_cartesian(xlim = c(0.73, 0.9),
                  ylim = c(70, 5))
ggsave("drafts/missing/figures/subset_hybrid_grayscale.png",
       height = 5, width = 5, scale = 1.25)

# In-text HDIs
results_subset %>% 
  mutate(discrim = round(discrim, 3),
         calib = round(calib, 2)) %>% 
  pivot_longer(3:4) %>% 
  group_by(type, name) %>%
  summarise(median = median(value),
    HDI = hdi(value, 0.95)) %>% 
  ungroup() 

results_hybrid %>% 
  mutate(discrim = round(discrim, 3),
         calib = round(calib, 2)) %>% 
  pivot_longer(3:4) %>% 
  group_by(type, name) %>%
  summarise(median = median(value),
    HDI = hdi(value, 0.95)) %>% 
  ungroup() 
```
